Split xcpclient into smaller modules
